Check existence before asking

+++ zip-content/scripts/install.sh
@@ -261,11 +261,11 @@ if test "${INSTALL_NEWPIPE:?}" -ne 0; then
   fi
 fi
 
-if test "${API}" -ge 23 && test "${live_setup_enabled:?}" = 'true'; then
+if test "${API}" -ge 23 && test "${live_setup_enabled:?}" = 'true' && test -f "${TMP_PATH}/files/variants/AndroidAuto.apk"; then
   choose 'Do you want to install Android Auto stub?' '+) Yes' '-) No'
   if test "$?" -eq 3; then INSTALL_ANDROID_AUTO='1'; else INSTALL_ANDROID_AUTO='0'; fi
 fi
-if test "${API}" -ge 23 && test "${INSTALL_ANDROID_AUTO:?}" -ne 0; then
+if test "${API}" -ge 23 && test "${INSTALL_ANDROID_AUTO:?}" -ne 0 && test -f "${TMP_PATH}/files/variants/AndroidAuto.apk"; then
   move_rename_file "${TMP_PATH}/files/variants/AndroidAuto.apk" "${TMP_PATH}/files/priv-app/AndroidAutoStubPrebuilt.apk"
 else
   delete "${TMP_PATH}/files/variants/AndroidAuto.apk"
